requirements.txt
django
django-heroku
djangorestframework
gunicorn
ipython
pillow
python-decouple
